+++ images/jupyter-user/bundle.sh
@@ -41,14 +41,14 @@ touch ${ARCHIVE_DIR}/pip.conf
 #Copy the orbit src code
 if [ -d "${DIR}/../aws-orbit/" ]; then
     rsync -av --progress ${DIR}/../aws-orbit/ ${ARCHIVE_DIR}/aws-orbit \
-    --exclude remotectl.out --exclude .orbit.out --exclude .mypy_cache \
+    --exclude codeseeder.out --exclude .orbit.out --exclude .mypy_cache \
     --exclude build --exclude dist --exclude docs \
     --exclude aws_orbit.egg-info --exclude __pycache__
 fi
 
 if [ -d "${DIR}/../aws-orbit-sdk/" ]; then
     rsync -av --progress ${DIR}/../aws-orbit-sdk/ ${ARCHIVE_DIR}/aws-orbit-sdk \
-    --exclude remotectl.out --exclude .orbit.out --exclude .mypy_cache \
+    --exclude codeseeder.out --exclude .orbit.out --exclude .mypy_cache \
     --exclude build --exclude dist --exclude docs \
     --exclude aws_orbit_sdk.egg-info --exclude __pycache__
 fi
@@ -58,7 +58,7 @@ if [ -d "${DIR}/../jupyterlab_orbit/" ]; then
     --exclude aws_orbit_jupyterlab_orbit.egg-info --exclude build \
     --exclude dist --exclude jupyterlab_orbit.egg-info \
     --exclude lib --exclude node_modules \
-    --exclude remotectl.out --exclude .orbit.out --exclude .mypy_cache \
+    --exclude codeseeder.out --exclude .orbit.out --exclude .mypy_cache \
     --exclude __pycache__
 fi